Create recipe to install kernel and set it as the default
This creates a new install recipe which will set kernel-osuosl as the default
kernel. Without this, the platform default kernel will automatically be set as
the new default if a new one has been installed. This will override
kernel-osuosl even though it has a higher version number.

This recipe also installs the kernel-osuosl package automatically which should
simplify using this cookbook in the future.